requirements.txt
Flask==2.3.2
Werkzeug~=2.3.6
numpy==1.25.0
pandas~=2.0.2
Pillow~=9.5.0
# torch, torchvision: We need cpu versions for Heroku. The CUDA packages are far too big (slug size
timm~=0.9.2
matplotlib~=3.7.1
torch~=2.0.1
torchvision~=0.15.2
